pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "qcpinn"
version = "2.0.0"
description = "Noise-Resilient Trainable Embeddings for Quantum Physics-Informed Neural Networks"
readme = "README_v2.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "pennylane>=0.35",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "scipy>=1.10",
]

[project.optional-dependencies]
hardware = [
    "pennylane-qiskit",
    "qiskit-ionq",
]
dev = [
    "pytest",
    "black",
    "ruff",
]
hpo = [
    "optuna>=3.0",
    "pandas>=1.5",
]

[project.scripts]
qcpinn = "qcpinn.run:main"

[tool.setuptools.packages.find]
include = ["qcpinn*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
